Files are organized, before launching
