+++ submarine-workbench/workbench-web-ng/src/app/pages/workbench/workbench.module.ts
@@ -41,7 +41,6 @@ import { EnvironmentComponent } from './environment/environment.component';
     WorkbenchComponent,
     HomeComponent,
     WorkspaceComponent,
-    ExperimentComponent,
     DataComponent,
     ModelComponent,
     EnvironmentComponent
